consolidate lib/api folders skill and skills
