+++ front/lib/api/skills/icon_suggestion.ts
