+++ front/pages/api/w/[wId]/builder/skills/suggestions.ts
@@ -4,8 +4,8 @@ import * as reporter from "io-ts-reporters";
 import type { NextApiRequest, NextApiResponse } from "next";
 
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
-import type { SkillDescriptionSuggestionInputs } from "@app/lib/api/skill/description_suggestion";
-import { getSkillDescriptionSuggestion } from "@app/lib/api/skill/description_suggestion";
+import type { SkillDescriptionSuggestionInputs } from "@app/lib/api/skills/description_suggestion";
+import { getSkillDescriptionSuggestion } from "@app/lib/api/skills/description_suggestion";
 import type { Authenticator } from "@app/lib/auth";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types";
